Corected KPI, match, load and unload services, Added migrations for delay rules

+++ Core/Entities/KPIData/KPI/Models/DB/KPI.cs
@@ -10,12 +10,12 @@ public partial class KPI : BaseEntity, IBaseEntity<KPI, DTOKPI>
 {
     public int NbCandidats { get; set; }
     public int Threshold { get; set; }
-    public double MScore { get; set; }
-    public double NMminScore { get; set; }
-    public double NMmaxScore { get; set; }
+    public int MScore { get; set; }
+    public int NMminScore { get; set; }
+    public int NMmaxScore { get; set; }
     public double NMAvg { get; set; }
     public double NMStdev {  get; set; }
-    public double ComputeTime {  get; set; }
+    public long ComputeTime {  get; set; }
 
     public MatchableCycle StationCycle { get; set; }
     public List<TenBestMatch> TenBestMatches { get; set; } = new();
